error message adjusted to work with Search and Random buttons
